Fix bug in Film::AddSpat().
(Due to the fact that the Camera importance methods return >0 importance within the sample bounds but then Film::AddSplat() just uses a box filter.)

With BDPT, the pixels along the left and top edges would receive ~2x the contributions they were supposed to if a pixel filter with > 1x1 extent was used.

The issue was that p would come in  with negative values in one of their coordinates between -1 and 0. The conversion from Point2f to Point2i would turn those to zero and then their contribution would be included, even though it wasn't supposed to be.