Fix Curve Log2 computation when Float=Double
Before, it would end up calling the 64-bit FloatToBits, then letting the result be truncated to 32-bits, and then interpreting some bits of that as exponent. For a double, it ended up using random bits from the significand, rather than the exponent.

Discovered while chasing down #221, though not the cause of it.