Remove pError scale in OffsetRayOrigin() when PBRT_FLOAT_AS_DOUBLE is enabled.
Before, pError was scaled by 1024 in that case; the idea was that there is
so much excess precision with 64-bit floats that it wouldn't matter to bump
up the error that much.

The problem with that is the Curve shape: because it has the behavior of
facing the incident ray, the ray epsilon has to be fairly large--half the
curve's width.

Multiplying that by 1024 in turn totally broke spawned rays, which
effectively wouldn't intersect anything...

Fixes issue #221.