Fix orientation of wh in MicrofacetReflection::f
Always orient wh to be on the same side of the surface as the normal; in turn, this means that the cos theta value passed to Fresnel::Evaluate() has the right sign.

This in particular fixes the handling of total internal reflection with this BxDF, both accounting for it when it should be and not incorrectly including it when there is none. (Rendered images may change due to this, especially noticeable in cases where TIR was incorrectly happening with non-transmissive surfaces, giving bright highlights.)

Fixes issues #214 and #254.